pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langfuse-pydantic-ai"
description = "langfuse-pydantic-ai"
keywords = ["langfuse", "pydantic-ai"]
requires-python = ">=3.9"
dependencies = ["pydantic-ai-slim", "langfuse"]
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-cov", "dirty-equals"]
docs = ["Sphinx", "pydata-sphinx-theme", "sphinx-click", "autodoc_pydantic"]

[[project.authors]]
name = "wh1isper"
email = "jizhongsheng957@gmail.com"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
text = "BSD 3-Clause License"

[project.urls]
Source = "https://github.com/wh1isper/langfuse_pydantic_ai"

[tool.check-manifest]
ignore = [".*"]

[tool.hatch.version]
path = "langfuse_pydantic_ai/__init__.py"
